Warn if scope is absent from credentials.
Autodesk's OAuth service will soon require clients to specify scope when
requesting an OAuth token. This code now looks for a scope specifier in
the same file as the client ID and secret. If the scope specifier does
not exist, a warning is presented to the user, and a default scope
specifier is used. The user can suppress the warning by adding a scope
specifier to the credentials file.